#211 remove `test_comments_count`
