+++ test/test__helper.rb
@@ -51,32 +51,3 @@ class Minitest::Test
     Fbe.github_graph(options: Judges::Options.new('testing' => true))
   end
 end
-
-class Fbe::Graph::Fake
-  def resolved_conversations(_owner, _name, _number)
-    [
-      {
-        'id' => 'PRRT_kwDOK2_4A85BHZAR',
-        'isResolved' => true,
-        'comments' =>
-        {
-          'nodes' =>
-          [
-            {
-              'id' => 'PRRC_kwDOK2_4A85l3obO',
-              'body' => 'first message',
-              'author' => { '__typename' => 'User', 'login' => 'reviewer' },
-              'createdAt' => '2024-08-08T09:41:46Z'
-            },
-            {
-              'id' => 'PRRC_kwDOK2_4A85l3yTp',
-              'body' => 'second message',
-              'author' => { '__typename' => 'User', 'login' => 'programmer' },
-              'createdAt' => '2024-08-08T10:01:55Z'
-            }
-          ]
-        }
-      }
-    ]
-  end
-end
